requirements.txt
aio_pika==6.8.2
python-dotenv==0.20.0
